services/server/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
flask-migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.4
# apscheduler # flaskの時は専用のものを遣う
# chromedriver-binary # google-chromeのバージョンに合わせてchromedriverのバージョンを会わないといけないで別で定義
flask_apscheduler
selenium
numpy
pandas
python-dotenv #.flaskenvを読み込む用